requirements.txt
soundcloud-lib
python-telegram-bot
python-dotenv